MyReview component renders on front end and retrieves sentiment results. Next step to POST sentiment results to our database/server
